Increase falling speed for trees and faster chunk detection
- Increased GravityScale from 2.0 to 5.0 for faster tree falling
- Check for floating chunks every 0.1s instead of 0.5s
- Increased angular damping to prevent excessive spinning

+++ src/player/components.rs
@@ -2,8 +2,8 @@ use bevy::prelude::*;
 
 #[derive(Resource)]
 pub struct Player {
-    pub x: i32,
-    pub y: i32,
+    pub x: f32,  // Use float for smooth sub-pixel movement
+    pub y: f32,
     pub vx: f32,
     pub vy: f32,
     pub width: i32,
@@ -13,8 +13,8 @@ pub struct Player {
 impl Default for Player {
     fn default() -> Self {
         Self {
-            x: 200,  // Spawn away from sand pile at x=400
-            y: 500,
+            x: 200.0,  // Spawn away from sand pile at x=400
+            y: 500.0,
             vx: 0.0,
             vy: 0.0,
             width: 8,
